backend/requirements.txt
fastapi[all]
psycopg2-binary
openai
uvicorn